refactor(domain): firebase 도메인 로직 분리 app 인증과 firestore 정보 분리
